+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@
 ## 2.0.0
 
 - New 404 page
+- New: See all the companies your citizens work at in one view (Citizen -> Manage employment status -> See it!)
+- Major: Tons of bug fixes
 - Major: Rewritten to [new framework](https://nextjs.org) to improve developer experience and performance
 - Major: able to have custom Open graph tags - TODO: add link here