+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 2.0.1
+
+- Minor: fixed not being able to create tow & taxi calls
+- Minor: updated [FiveM addons](https://github.com/Dev-CasperTheGhost/snaily-cadv3/tree/main/public/downloads)
+
 ## 2.0.0
 
 - New 404 page